Factor out settings panel
Establish ability to have modular UI sections
